#18: Restructuring cli code (commands & flags)
