+++ cmd/atlas/flag/clean.go
