--- cmd/docker.go
@@ -1,23 +1,0 @@
-package cmd
-
-import (
-	"context"
-
-	"github.com/ahstn/atlas/pkg/docker"
-	"github.com/urfave/cli"
-)
-
-// Docker defines the command for the cli and the logic to utilise Docker
-var Docker = cli.Command{
-	Name:    "docker",
-	Aliases: []string{"d"},
-	Usage:   "build an application's Dockerfile",
-	Action:  DockerAction,
-}
-
-// DockerAction handles building a container
-func DockerAction(c *cli.Context) error {
-	ctx := context.Background()
-
-	return docker.ImageBuild(ctx)
-}
